pyproject.toml
[tool.poetry]
name = "quiz-compiler"
version = "1.0.0"
description = "Compiler for quizes"
authors = ["Andrii Gerasymchuk <gera3dartist@gmail.com>"]
readme = "README.md"
packages = [{include = "quiz_compiler"}]

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.86.0"
click = "^8.1.3"
oauth2client = "^4.1.3"
dependency-injector = "^4.41.0"
pillow = "^9.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
